+++ R/get_coord_from_string.R
@@ -1,5 +1,34 @@
-get_coord_from_string <- function(x , pattern = "Longitude") {
+#' Getting coordinates from a string provided by Ogimet web portal
+#'
+#' Internal function for cleaning coordinates' metadata provided by Ogimet
+#' @param txt string element with coordinates from Ogimet
+#' @param pattern which element (longitude or latitude) to extract
+#'
+#' @format The returned object is the geographic coordinates using WGS84 (EPSG:4326) in decimal format.
+#' Negative values mean western or southern Hemisphere
+#'
+#' @examples
+#' \donttest{
+#'  txt <- "12330:   Poznan (Poland)\nLatitude: 52-25N    Longitude: 016-50E    Altitude: 86 m."
+#'   get_coord_from_string(txt, pattern = "Latitude")
+#' }
+#'
+
+
+
+get_coord_from_string <- function(txt , pattern = "Longitude") {
   tt <- gregexpr(pattern, txt)
   start <- tt[[1]][1] + attributes(tt[[1]])$match.length + 1
-  substr(txt, start = start, stop = start + 8)
+  tmp <- trimws(substr(txt, start = start, stop = start + 8))
+  tmp <- strsplit(tmp, "-")[[1]]
+  hemisphere <- gsub("[0-9]", "", strsplit(tmp, "-")[2])
+  tmp <- gsub("[A-Z]", "", strsplit(tmp, "-"))
+
+  wsp <- as.numeric(tmp)[1]+(as.numeric(tmp)[2]*1.6666)/100
+
+  if( hemisphere %in% c("W","S") ) {
+    wsp <- wsp*-1
+  }
+return(wsp)
 }
+